Template some more form creation over type
